Extract S3 backend from storage module
